Moved implementation to utils.
